Organizing the intent files a little better
